ref: Extracts an AbstractParticipant class,
which is to be used for the bridge-to-bridge (Octo) channels.
